Add web-based UI for ElevenLabs Soundboard
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Anguscrane/Elevenlabs-Soundboard?shareId=XXXX-XXXX-XXXX-XXXX).
