--- ElevenLabs.bat
@@ -1,3 +1,0 @@
-@echo off
-python "%~dp0ElevenLabs Ai Soundboard.py"
-pausepause
